Rename database/database.c => database/common.c
