+++ database/common.h
